fix(ci): remove benchmark tool and deduplicate conditional-write tests for SonarCloud gate
SonarCloud Quality Gate on PR #11541 (Phase-1 conditional writes) failed
on two new-code conditions:

1. 5 Security Hotspots — all math/rand "weak PRNG" in
   test/benchmark/conditional_writes/main.go (lines ~291, 411, 417,
   420, 427). The benchmark is a dev/load-testing CLI whose results are
   already recorded; it is not part of the Phase-1 feature.

2. 8.0% code duplication on new code (gate requires ≤3%). The five
   conditional test files repeated storobj builders, HTTP insert helpers,
   schema setup, and the waitForSchemaOnAllNodes poll loop.

Fix 1 — Remove benchmark load tool:
  - git rm test/benchmark/conditional_writes/ (main.go + README.md)
  - Remove the now-dangling exclusion line from tools/linter_go_routines.sh

Fix 2 — Deduplicate test helpers:
  adapters/repos/db/:
  - New file conditional_write_test_helpers_test.go consolidates
    buildConditionalObject (OnlyIfNotExists), buildOnlyIfExistsObject
    (OnlyIfExists), buildUnconditionalObject (no condition),
    buildMTConditionalObject (MT + tenant), and isPreconditionFailed.
  - shard_write_put_concurrent_test.go: remove the now-duplicate
    buildConditionalObject definition (was also there).
  - conditional_batch_mixed_test.go: remove newConditionalObj,
    newOnlyIfExistsObj, newUnconditionalObj, isPreconditionFailed; rename
    callers to canonical names from shared file.
  - conditional_durability_test.go: remove newConditionalInsertObj;
    rename callers.
  - conditional_multitenant_test.go: remove buildMTConditionalObject
    (now in shared file).

  test/acceptance/conditional_writes/:
  - New file helpers_test.go consolidates insertObjectHTTP (generic
    conditional/unconditional), condInsertHTTP, uncondInsertHTTP,
    setupRF3Class, and waitForSchemaOnAllNodes.
  - production_readiness_multinode_test.go: remove prodCondInsertHTTP,
    prodUncondInsertHTTP, setupProdClass, waitForSchemaOnAllNodes (~170
    lines); update callers to shared helpers.
  - ha_regression_test.go: remove conditionalInsertHTTP,
    unconditionalInsertHTTP, setupHAClass (~100 lines); update call
    sites to shared helpers (strfmt.UUID cast to string at each site).

Net: 444 lines removed from existing files; 265 lines added across two
shared helper files; 1000 lines of benchmark code deleted.

go build ./... clean. go vet pre-existing warnings only (dot_arm64.s,
helper_for_test.go:216 context leak — both present before this change).
go test -c ./test/acceptance/conditional_writes/... and
-tags integrationTest ./adapters/repos/db/... both compile clean.
Unit tests (TestConditional*, TestInsertIfNotExists*, TestVersion*) green.
./tools/linter_go_routines.sh passes.
